hmb: add fixture and examples
