+++ scripts/test_examples.py
@@ -743,3 +743,48 @@ def test_reset_time(pcie):
     subsystem.poweron()
     nvme0.reset()
     
+
+def test_controller_with_hmb(nvme0n1, hmb):
+    for i in range(10):
+        logging.info(i)
+        with nvme0n1.ioworker(io_size=8,
+                              lba_random=False,
+                              qdepth=8,
+                              read_percentage=0,
+                              time=3):
+            pass
+
+        
+def test_controller_reset_with_hmb(nvme0, nvme0n1, buf):
+    nvme0n1.format(512)
+
+    hmb_size = nvme0.id_data(275, 272)
+    hmb_buf = d.Buffer(4096*hmb_size)
+    assert hmb_buf
+    hmb_list_buf = d.Buffer(4096)
+    assert hmb_list_buf
+    hmb_list_buf[0:8] = hmb_buf.phys_addr.to_bytes(8, 'little')
+    hmb_list_buf[8:12] = hmb_size.to_bytes(4, 'little')
+    hmb_list_phys = hmb_list_buf.phys_addr
+    
+    for i in range(10):
+        logging.info(i)
+        with nvme0n1.ioworker(io_size=8,
+                              lba_random=False,
+                              qdepth=8,
+                              read_percentage=0,
+                              time=30):
+            nvme0.setfeatures(0x0d,
+                              cdw11=1,
+                              cdw12=hmb_size,
+                              cdw13=hmb_list_phys&0xffffffff,
+                              cdw14=hmb_list_phys>>32,
+                              cdw15=1).waitdone()
+            time.sleep(5)
+            nvme0.setfeatures(0x0d, cdw11=0).waitdone()
+            nvme0.reset()
+
+    logging.info(hmb_buf.dump(64*1024))
+    del hmb_buf
+    del hmb_list_buf
+
